Support working paths with spaces
Added a new function for parsing out the working path from the tree strings that are output by 'git worktree list'.
Used the new function to set the working directory instead of relying on positional parameters set by 'git worktree list'.
Configured 'awk' calls to use tabs instead of spaces as field separators for a line.
Distinguished the console message that is printed when switching to the main worktree.